resources/requirements-py3.txt
requests
capitains_nautilus>=0.0.5
flask_nemo>=1.0.0b1
nemo_oauth_plugin>=0.0.1
flask_oauthlib>=0.9.3
gunicorn
